defer duplicate Listener functionality to a generic Listener
